backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1

sqlalchemy>=2.0.25
pyodbc==5.1.0

python-dotenv==1.0.0
requests==2.31.0

motor>=3.3.2
beanie>=1.25.0
redis==5.0.1

langgraph>=0.0.40
langchain>=0.1.16
langchain-groq>=0.1.0
langchain-community>=0.0.38
langchain-huggingface>=0.0.3
pydantic>=2.0.0
sentence-transformers>=2.2.2

google-generativeai>=0.3.2

twilio==8.10.0
apscheduler==3.10.4

Pillow>=10.4.0

gunicorn==21.2.0
celery>=5.3.6
tenacity>=8.2.3